items: disable the delete action if there is a fee
Warning: the PatronTransaction data must be reindexed.

* Adds the item pid when indexing the PatronTransaction record.
* Closes #2102
