+++ rero_ils/modules/patron_transactions/mappings/v7/patron_transactions/patron_transaction-v0.0.1.json
@@ -81,6 +81,16 @@
           }
         }
       },
+      "item": {
+        "properties": {
+          "type": {
+            "type": "keyword"
+          },
+          "pid": {
+            "type": "keyword"
+          }
+        }
+      },
       "total_amount": {
         "type": "float"
       },
